pyproject.toml
[project]
name = "gygax"
version = "0.1.0"
description = "Fiddling with AI and DnD"
authors = [
    {name = "Eric Ekholm",email = "eric.ekholm@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "chromadb (>=1.0.0,<2.0.0)",
    "google-genai (>=1.9.0,<2.0.0)",
    "pysqlite3-binary (>=0.5.4,<0.6.0)",
    "google-api-core (>=2.24.2,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

